+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 - The progress reporter will automatically be stopped after 10 seconds, if it's still running. This
   ensures that it is stopped even when the connection to PostgreSQL fails.
 
+- Further workarounds for semi-compatible PostgreSQL variants QuestDB and CrateDB.
+
 
 ## [0.21] - 2025-01-12